requirements.txt
influxdb-client
python-dotenv
psutil
getmac
apscheduler